data_producer/requirements.txt
pyspark
confluent_kafka
fastavro
